refactor(panel): extract app-update prompts and Send Feedback into composables
Pure relocation. PanelApp.vue 1203 -> 1118 lines.

- useAppUpdatePrompts: showAppUpdateRestartPrompt, showAppUpdateDownloadPrompt
- useSendFeedback: app version cache + onOpenFeedback subscription + click handler

Amp-Thread-ID: https://ampcode.com/threads/T-019e2311-6aeb-7349-9fd9-560288cfbe38
